fmt and refactor thingy
