+++ src-tauri/src/plugin_management/download.rs
